+++ llvm/include/llvm/ExecutionEngine/Orc/RTBridge/SPS/MemoryAccessProxySpecs.h
@@ -1,4 +1,4 @@
-//===-------- ProxySpecs.h - SPS-based Call Wrappers ------------*- C++ -*-===//
+//===-- MemoryAccessProxySpecs.h - SPS specs for mem access -----*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,108 +6,21 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// SPS-based implementations of the RTBridge proxy interfaces.
-//
-// These implement the rt::Proxy interfaces by invoking executor-side wrapper
-// functions in the runtime's controller interface, using Simple Packed
-// Serialization to encode arguments and decode results.
+// SPS ProxySpecs (signatures, controller-interface names, and dispatch) for the
+// MemoryAccessProxies.
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_EXECUTIONENGINE_ORC_RTBRIDGE_SPS_PROXYSPECS_H
-#define LLVM_EXECUTIONENGINE_ORC_RTBRIDGE_SPS_PROXYSPECS_H
+#ifndef LLVM_EXECUTIONENGINE_ORC_RTBRIDGE_SPS_MEMORYACCESSPROXYSPECS_H
+#define LLVM_EXECUTIONENGINE_ORC_RTBRIDGE_SPS_MEMORYACCESSPROXYSPECS_H
 
-#include "llvm/ExecutionEngine/Orc/Core.h"
-#include "llvm/ExecutionEngine/Orc/RTBridge/Proxy.h"
+#include "llvm/ExecutionEngine/Orc/RTBridge/MemoryAccessProxies.h"
+#include "llvm/ExecutionEngine/Orc/RTBridge/SPS/ProxySpec.h"
 #include "llvm/ExecutionEngine/Orc/Shared/TargetProcessControlTypes.h"
 
 #include <cstdint>
 
 namespace llvm::orc::rt::sps {
-
-template <typename ProxyT, typename SPSSigT, const char *DefaultName,
-          typename FnType = typename ProxyT::FnType>
-class ProxySpec;
-
-template <typename ProxyT, typename SPSSigT, const char *DefaultName,
-          typename RetT, typename... ArgTs>
-class ProxySpec<ProxyT, SPSSigT, DefaultName, RetT(ArgTs...)> {
-
-  using CalleeRetT = typename ProxyT::CalleeRetT;
-  using ErrorRetT = typename ProxyT::ErrorRetT;
-
-  static void consumeResult(Error &Err) { consumeError(std::move(Err)); }
-
-  template <typename T> static void consumeResult(T &V) {}
-
-  template <typename T> static void consumeResult(Expected<T> &E) {
-    consumeError(E.takeError());
-  }
-
-public:
-  static constexpr const char *Name = DefaultName;
-
-  static void dispatch(unique_function<void(ErrorRetT)> OnComplete,
-                       ExecutionSession &ES, ExecutorAddr CalleeAddr,
-                       const ArgTs &...Args) {
-    if constexpr (std::is_void_v<CalleeRetT>) {
-      // Void result: the executor-side function produces no value, so the only
-      // thing to report is the dispatch error (success if the call ran).
-      ES.callSPSWrapperAsync<SPSSigT>(CalleeAddr, std::move(OnComplete),
-                                      Args...);
-    } else {
-      ES.callSPSWrapperAsync<SPSSigT>(
-          CalleeAddr,
-          [OnComplete = std::move(OnComplete)](Error SerErr,
-                                               CalleeRetT Result) mutable {
-            if (SerErr) {
-              consumeResult(Result);
-              return OnComplete(std::move(SerErr));
-            }
-            // For an Error/Expected callee this forwards the callee's own
-            // result; for a plain value it is wrapped into Expected.
-            return OnComplete(std::move(Result));
-          },
-          Args...);
-    }
-  }
-};
-
-using CallMainSPSSig = int64_t(shared::SPSExecutorAddr,
-                               shared::SPSSequence<shared::SPSString>);
-inline constexpr char CallMainCIName[] = "orc_rt_ci_sps_call_main";
-/// SPS proxy for rt::CallMainProxy: runs a main-like function
-/// (int(int argc, char *argv[])) in the executor.
-using CallMainProxySpec =
-    ProxySpec<rt::CallMainProxy, CallMainSPSSig, CallMainCIName>;
-
-using CallVoidVoidSPSSig = void(shared::SPSExecutorAddr);
-inline constexpr char CallVoidVoidCIName[] = "orc_rt_ci_sps_call_void_void";
-/// SPS proxy for rt::CallVoidVoidProxy: runs a void() function in the executor.
-/// WARNING: This Proxy is experimental and may be removed.
-using CallVoidVoidProxySpec =
-    ProxySpec<rt::CallVoidVoidProxy, CallVoidVoidSPSSig, CallVoidVoidCIName>;
-
-using CallInt32VoidSPSSig = int32_t(shared::SPSExecutorAddr);
-inline constexpr char CallInt32VoidCIName[] = "orc_rt_ci_sps_call_int32_void";
-/// SPS proxy for rt::CallInt32VoidProxy: runs an int32_t() function in the
-/// executor.
-/// WARNING: This Proxy is experimental and may be removed.
-using CallInt32VoidProxySpec =
-    ProxySpec<rt::CallInt32VoidProxy, CallInt32VoidSPSSig, CallInt32VoidCIName>;
-
-using CallInt32Int32SPSSig = int32_t(shared::SPSExecutorAddr, int32_t);
-inline constexpr char CallInt32Int32CIName[] = "orc_rt_ci_sps_call_int32_int32";
-/// SPS proxy for rt::CallInt32Int32Proxy: runs an int32_t(int32_t) function in
-/// the executor.
-/// WARNING: This Proxy is experimental and may be removed.
-using CallInt32Int32ProxySpec =
-    ProxySpec<rt::CallInt32Int32Proxy, CallInt32Int32SPSSig,
-              CallInt32Int32CIName>;
-
-// Memory-access proxies. Unlike the Call* proxies above, these target wrappers
-// that perform the operation directly, so they take the operation's data
-// arguments and no callee address.
 
 using MemWriteUInt8sSPSSig =
     void(shared::SPSSequence<shared::SPSMemoryAccessUInt8Write>);
@@ -207,4 +120,4 @@ using MemReadStringsProxySpec =
 
 } // namespace llvm::orc::rt::sps
 
-#endif // LLVM_EXECUTIONENGINE_ORC_RTBRIDGE_SPS_PROXYSPECS_H
+#endif // LLVM_EXECUTIONENGINE_ORC_RTBRIDGE_SPS_MEMORYACCESSPROXYSPECS_H
